pyproject.toml
[project]
name = "pingmebot"
version = "0.1.0"
description = ""
authors = [
    {name = "latamrisus",email = "swag@swag.swag"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.124.2,<0.125.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic (>=2.0.0)",
    "pydantic-settings (>=2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "alembic (>=1.17.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
